Add files via upload
Given a number n, for each integer i in the range¶
from 1 to n inclusive, print one value per line as follows: • Ifi is a multiple of both 3 and 5, print FizzBuzz. • If iis a multiple of 3(but not 5), print Fizz. • Ifi is a multiple of 5(but not 3), print Buzz. • If i is not a multiple of 3 or 5, print the value of i.​